requirements.txt
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.4.26
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.4.2
constantly==23.10.4
cryptography==45.0.4
daphne==4.1.1
Django==5.2.3
django-allauth==65.9.0
django-environ==0.12.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.1
oauthlib==3.2.2
pillow==11.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.1.0
redis==5.0.4
requests==2.32.4
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools==80.9.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==23.1.1
typing_extensions==4.14.0
urllib3==2.4.0
whitenoise==6.9.0
zope.interface==7.2
